+++ src/mswindows.py
@@ -24,10 +24,8 @@ THE SOFTWARE.
 
 """ Windows specific module """
 
-import io
-import ctypes
-from PIL import Image
 import platform
+
 
 def windows():
     """ checking: system Windows or normal OS """
@@ -36,35 +34,6 @@ def windows():
     else:
         result = 0
     return result
-
-
-def windows_copy_to_clipboard(file_in):
-    """
-    https://stackoverflow.com/questions/21319486/copy-pil-pillow-image-to-windows-clipboard
-    """
-
-    msvcrt = ctypes.cdll.msvcrt
-    kernel32 = ctypes.windll.kernel32
-    user32 = ctypes.windll.user32
-
-
-    img = Image.open(file_in)
-    output = io.BytesIO()
-    img.convert('RGB').save(output, 'BMP')
-    data = output.getvalue()[14:]
-    output.close()
-
-    CF_DIB = 8
-    GMEM_MOVEABLE = 0x0002
-
-    global_mem = kernel32.GlobalAlloc(GMEM_MOVEABLE, len(data))
-    global_data = kernel32.GlobalLock(global_mem)
-    msvcrt.memcpy(ctypes.c_char_p(global_data), data, len(data))
-    kernel32.GlobalUnlock(global_mem)
-    user32.OpenClipboard(None)
-    user32.EmptyClipboard()
-    user32.SetClipboardData(CF_DIB, global_mem)
-    user32.CloseClipboard()
 
 
 # EOF